Added computer vision functionality to the robot
Games can now be played with a whiteboard marker against the robotic arm. OpenCV has been used with a webcam fixed to the robot arm such that the program can watch the user's moves drawn and respond to them according to the game. Computer vision techniques such as Morphology, Blob Detection and Foreground Detection (a form of motion detection) have been applied.